:truck: [#99] Move fixture to confest.py for generic usage
